Extract atom-info footer to its own component Removes unused function in atom-utils
