+++ webofneeds/won-owner-webapp/src/main/webapp/style/_atom-info.scss
@@ -23,121 +23,12 @@ won-atom-info {
     display: grid;
   }
 
-  .atom-info__footer {
+  > won-atom-footer {
     grid-area: footer;
     display: grid;
     grid-gap: 0.5rem;
     padding: 0.5rem;
     box-sizing: border-box;
     border-top: $thinGrayBorder;
-
-    &__button {
-      width: 100%;
-
-      > svg.won-button-icon {
-        --local-primary: white;
-      }
-    }
-
-    & > *:not(:last-child) {
-      // a bit of space between elements in the footer
-      margin-bottom: 0.5rem;
-    }
-
-    &__infolabel {
-      text-align: center;
-    }
-
-    &__header {
-      font-size: $normalFontSize;
-      font-weight: bold;
-      color: $won-primary-text-color;
-    }
-    &__matches {
-      &__header {
-        font-size: $smallFontSize;
-        color: $won-primary-text-color;
-        font-weight: lighter;
-      }
-
-      &__list {
-        display: grid;
-        grid-gap: 1rem;
-        padding: 0.5rem 0;
-
-        grid-template-columns: repeat(auto-fill, minmax(16rem, 1fr));
-
-        > .atom-info__footer__adhocbutton,
-        > won-atom-header {
-          padding: 0.5rem;
-          border: $thinGrayBorder;
-
-          &:hover {
-            background: $won-lighter-gray;
-            cursor: pointer;
-          }
-        }
-      }
-    }
-    &__adhocbutton {
-      display: grid;
-      grid-template-areas: "icon main";
-      grid-template-columns: min-content 1fr;
-      grid-column-gap: 0.5rem;
-      min-width: 0;
-
-      &__icon {
-        background: $won-primary-color;
-        box-sizing: border-box;
-        padding: 0.25rem;
-        @include fixed-square($postIconSize);
-
-        &__svg {
-          --local-primary: #{$won-secondary-text-color};
-          height: 100%;
-          width: 100%;
-        }
-      }
-
-      &__right {
-        grid-area: main;
-        display: grid;
-        grid-template-areas: "topline" "subtitle";
-        min-width: 0;
-
-        &__topline {
-          grid-area: topline;
-          min-width: 0;
-
-          &__notitle {
-            min-width: 0;
-            white-space: nowrap;
-            text-overflow: ellipsis;
-            overflow: hidden;
-            font-weight: 400;
-            color: $won-subtitle-gray;
-          }
-        }
-
-        &__subtitle {
-          grid-area: subtitle;
-          display: grid;
-          grid-template-columns: 1fr min-content;
-          color: $won-subtitle-gray;
-          font-size: $smallFontSize;
-          min-width: 0;
-
-          &__type {
-            text-overflow: ellipsis;
-            overflow: hidden;
-            white-space: nowrap;
-          }
-        }
-      }
-
-      > svg.won-button-icon {
-        --local-primary: white;
-      }
-    }
   }
 }
